requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Web framework
streamlit>=1.28.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
fdb>=2.0.2

# Vector search
faiss-cpu>=1.7.4

# LLM utilities
tiktoken>=0.5.0

# Local ML dependencies removed - using cloud APIs only

# LangChain SQL Agent Tools
langchain-experimental>=0.0.40
sqlalchemy>=2.0.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0
networkx>=3.0
sqlglot>=25.0.0

# AI Observability
arize-phoenix>=1.0.0

# LangGraph for advanced workflows  
langgraph>=0.4.8

# SQL parsing and validation
sqlglot>=26.25.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
responses>=0.23.0
flake8>=5.0.0
black>=22.0.0
isort>=5.11.0
bandit>=1.7.0
pre-commit>=3.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0